refactors getRangeProgressData Progress model function
